hotfix/#37: Added LoginRequiredMixin to the  OrderRetrieveCancelView
